Separate procedures.py into submodules & add execute
